requirements.txt
Django==5.2.8
django-allauth==65.13.0
stripe==13.2.0
python-dotenv==1.2.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
APScheduler==3.10.4
pytz==2024.1
